adjust RomajiToKatakana
